I made the following updates to the character views
Index: Added a for loop to dynamically add the name and picture of each character in the database.  Added logic to create a new row after every four characters.  Added link to view detail if you click on character name or picture.

Detail: Added image and changed the display of the detail.  Added links here to delete character and edit character.

Edit: Changed the 'Save' button to redirect to the detail view instead of the index view.